+++ ggml/src/ggml-openvino/ggml-openvino-extra.h
@@ -104,14 +104,6 @@ bool ggml_openvino_release_weights_enabled(const std::string & device);
 // Check if running on NPU
 bool ggml_openvino_is_npu();
 
-// Host weight-buffer release (GGML_OPENVINO_RELEASE_WEIGHTS, GPU only).
-// register: record a host weight buffer (idempotent per data pointer).
-// release:  madvise(MADV_DONTNEED) all registered buffers, dropping their RSS.
-// released: true once release has run (used to fail-fast on post-release recompile).
-void ggml_openvino_register_weight_buffer(void * data, size_t size);
-void ggml_openvino_release_weight_buffers();
-bool ggml_openvino_weight_buffers_released();
-
 // Get requantization type for a tensor type (returns nullopt if no requant needed)
 std::optional<ExtraQuantType> ggml_openvino_get_requant_type(const ggml_tensor * tensor, bool no_requant = false);
